Make button utility classes override more specific rules (e.g. in the Navbar menu)
These utility classes are overridden by more specific rules for anchors in the Navbar menu, particularly when used for a Navbar Button. The alternative, for now, is to include specific selectors for each of these classes to permit Navbar buttons to format as intended. Although !important is very undesirable, it may be somewhat tolerable for a utility class like this.