Adjust Navbar Button to use <button> element
Because Bootstrap v3 doesn't support using button classes on anchors inside navbar-nav, it looks like we will have to use <button> instead of <a>. This also means we have to use Javascript to enable redirects, and we should also use Javascript to insert the Navbar button on a page. This ensures the button is only rendered when Javascript is enabled on the client device.

KSS docs updated with info about how to use Navbar buttons.